Added in API.
The database is functioning well for a readonly database.
You can make elements in a json file
Or you can make them programatically.
I have added a Graph class that holds all information about the graph as a
whole, including the lists of Nodes and Edges.
A web server starts on port 8080.
You can query /showAllNodes, /showAllEdges, /showNode?id=, /showEdge?id=

Any processing of the database should currently happen before the webserver
has started, or could be added as an API call. Threading not currently
implemented - which would allow the database to do two or more things at
a time.